Move components of glossary dir to components
Move components of glossary directory that must not be
accessed by url to src/components/glossary
