+++ src/pages/glossary/index.tsx
@@ -9,10 +9,10 @@ import {
   Container,
 } from "@chakra-ui/react";
 import { ResultBox } from "../search/result-box";
-import { NewTermRequest } from "./new-term-request";
-import { ScrollToTopButton } from "./scroll-to-top-button";
+import { NewTermRequest } from "../../components/glossary/new-term-request";
+import { ScrollToTopButton } from "../../components/glossary/scroll-to-top-button";
 import { useTranslation } from "react-i18next";
-import { GlossaryDataFetchError } from "./glossary-data-fetch-error";
+import { GlossaryDataFetchError } from "../../components/glossary/glossary-data-fetch-error";
 
 const AllTerms = ({ terms }: any): JSX.Element => {
   const { t } = useTranslation();
